% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdConstruct.pdInd}
\alias{pdConstruct.pdInd}
\title{Construct pdInd object}
\usage{
\method{pdConstruct}{pdInd}(object, value = numeric(0),
  form = formula(object), nam = Names(object), data = sys.parent(),
  cov = NULL, zero = NULL, ...)
}
\arguments{
\item{object}{an object inheriting from the class \code{pdInd}, representing a positive definite matrix with zero covariances except in the first row and column.}
}
\description{
This function is a constructor for a \link{\code{pdInd}} object.
}

