% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdInd}
\alias{pdInd}
\title{Constuctor for Positive-Definite Matrix With Zero Covariances Between Predictor Random Effects}
\usage{
pdInd(value = numeric(0), form = NULL, nam = NULL, data = sys.parent(),
  cov = NULL)
}
\description{
TODO: add block diagonal covariances for, e.g. 3+-level factors.
}
\details{
This function is a constructor for the \code{pdInd} class, representing a positive-definite matrix with zero covariances except possibly in the first row and column. If the matrix associated with \code{object} is of dimension $n$, it is represented by $n + (n-1)$ unrestricted parameters representing a lower-triangular log-Cholesky decomposition. The first $n$ parameters are the logs of the diagonal elements of the matrix and the last $n-1$ components are the $n-1$ remaining elements of the lower-triangular decomposition corresponding the to the possibly non-zero covariances in the first row.
}

