\name{Tcells-package}
\alias{Tcells-package}
\docType{package}
\title{
Tcells: A package for Tcell dynamics with mixed models
}
\description{
For more information see \code{\link{Tcells}}. For some theory behind the 
packages use the command \code{vignette("Tcells")}
}
\details{
\tabular{ll}{
Package: \tab Tcells\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-06\cr
License: \tab GPL\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
