% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tcells.R
\name{Tcells}
\alias{Tcells}
\title{Tcells: an R package to estimate cell dynamics with mixed models.}
\usage{
Tcells()
}
\description{
Collection of pdMat, varClasses and corClasses objects and function to
fit cell dynamic models with mixed models using \code{\link[nlme]{lme}} in the package 'nlme'.
}
\seealso{
\url{http://127.0.0.1:11398/help/library/Tcells/doc/Plans.html}

\url{http://127.0.0.1/help/library/Tcells/doc/Plans.html}
To see a paper with some of the theory behind this package, use the command:\cr
   \code{vignette("Tcells")}
}

